# [`Optimization`]: Make murmur32 appendable and faster.