# Description

Implement`NonCryptographicHashAlgorithm` for `Murmure32`.
And make it faster.

Before:
```
BenchmarkDotNet v0.14.0, macOS Sequoia 15.1.1 (24B2091) [Darwin 24.1.0]
Apple M4 Pro, 1 CPU, 14 logical and 14 physical cores
.NET SDK 9.0.102
  [Host]     : .NET 9.0.1 (9.0.124.61010), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 9.0.1 (9.0.124.61010), Arm64 RyuJIT AdvSIMD


| Method                | Mean       | Error     | StdDev    |
|---------------------- |-----------:|----------:|----------:|
| Murmur32_HashToUInt32 | 138.497 us | 0.8307 us | 0.7771 us |
```

After:
```
BenchmarkDotNet v0.14.0, macOS Sequoia 15.1.1 (24B2091) [Darwin 24.1.0]
Apple M4 Pro, 1 CPU, 14 logical and 14 physical cores
.NET SDK 9.0.102
  [Host]     : .NET 9.0.1 (9.0.124.61010), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 9.0.1 (9.0.124.61010), Arm64 RyuJIT AdvSIMD


| Method                | Mean       | Error     | StdDev    |
|---------------------- |-----------:|----------:|----------:|
| Murmur32_HashToUInt32 |  89.359 us | 0.1607 us | 0.1424 us |
```

## Type of change

<!-- Please delete options that are not relevant. -->

- [x] Optimization (the change is only an optimization)
- [ ] Style (the change is only a code style for better maintenance or standard purpose)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules